+++ src/OpenTelemetry.OpAmp.Client/Internal/Listeners/Messages/ErrorResponseMessage.cs
@@ -5,7 +5,7 @@ using OpAmp.Proto.V1;
 
 namespace OpenTelemetry.OpAmp.Client.Internal.Listeners.Messages;
 
-internal class ErrorResponseMessage : IOpAmpMessage
+internal class ErrorResponseMessage : OpAmpMessage
 {
     public ErrorResponseMessage(ServerErrorResponse serverErrorResponse)
     {