+++ src/OpenTelemetry.OpAmp.Client/Internal/Listeners/Messages/RemoteConfigMessage.cs
@@ -5,7 +5,7 @@ using OpAmp.Proto.V1;
 
 namespace OpenTelemetry.OpAmp.Client.Internal.Listeners.Messages;
 
-internal class RemoteConfigMessage : IOpAmpMessage
+internal class RemoteConfigMessage : OpAmpMessage
 {
     public RemoteConfigMessage(AgentRemoteConfig agentRemoteConfig)
     {